Add html writer and remove glibc hash
Remove use of glibc hash function, and the GPLv3 license. The purpose
of having GPLv3 was to be able to use the glibc hash function more
extensively. Now when usage of the hash function is removed, the use
of GPLv3 is not needed any more.

Add a simple html-writer with appurtenant style.css.

COPYING is removed because it contained the license.

A tag has been added to add_feed(). The tag allows you to specify a
short name for the feed.

Indentation is updated to use tabs instead of four spaces.

Signed-off-by: anton <haglundanton+github@gmail.com>